Income endpoints & Table modification

+++ api/migrations/003_expense_table.py
@@ -4,7 +4,7 @@ steps = [
         """
         CREATE TABLE expenses (
             id SERIAL PRIMARY KEY NOT NULL,
-            expense_amount FLOAT NOT NULL,
+            expense_amount NUMERIC(10,2) NOT NULL,
             date DATE NOT NULL,
             description TEXT,
             user_id INTEGER NOT NULL,
